CDH-14778 - Backport HIVE-5454 - HCatalog runs a partition listing with an empty filter
When a table has a large number of partitions a query without a filter will timeout
getting the list of partitions. The issue is that in HCatInputFormat, there is first
a call to setInput and then setFilter. Both of which call InitializeInput.setInput.

This first call is executed without a filter and times out. In addition to this fix
the upstream patch HIVE-5454 removes methods which accept an InputJobInfo as that is
now a private object (depcrecated in Hcatalog 0.5 and Hive 0.11).

This backport changes the primary setInput(InputJobInfo) methods to perform only
one MS query for partitions and to use the partition filter on the first call.

To support this and transition to the new methods, we add the new method signatures
as well.